Fix shiftL bug that grew the bytestring size
When shifting a bytestring a number of bits larger than the number of bits in the actual bytestring, shiftL would originally grow the bytestring size to a length equal to the number of bytes needed to represent the final shifted value. This is not the intended behavior, shifting an array of bits should yield an array of bits the same size as the one put in. Thus, bits that are shifted out of bounds should be discarded.